Implement pr review changes requested
