+++ plugin-hrm-form/src/components/profile/ProfileContacts.tsx
@@ -17,7 +17,7 @@ import React from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 
 import { getPermissionsForContact, PermissionActions } from '../../permissions';
-import { Contact, RouterTask, Profile } from '../../types/types';
+import { Contact } from '../../types/types';
 import ContactPreview from '../search/ContactPreview';
 import * as ProfileTypes from '../../states/profile/types';
 import * as RoutingActions from '../../states/routing/actions';
