+++ plugin-hrm-form/src/components/search/ContactPreview/index.tsx
@@ -20,8 +20,7 @@ import { callTypes } from 'hrm-form-definitions';
 
 import ContactHeader from './ContactHeader';
 import TagsAndCounselor from '../TagsAndCounselor';
-import { PreviewWrapper } from '../../../styles/search';
-import { Flex } from '../../../styles/HrmStyles';
+import { Flex, PreviewWrapper } from '../../../styles/HrmStyles';
 import { ContactRawJson, Contact } from '../../../types/types';
 import { PreviewDescription } from '../PreviewDescription';
 import { isNonDataCallType } from '../../../states/validationRules';
